+++ client_src/js/printer-manager/printer-data.js
@@ -107,44 +107,30 @@ function corsWarningCheck(printer) {
 
 function checkForOctoPrintUpdate(printer) {
   let updateButton = document.getElementById(`octoprintUpdate-${printer._id}`);
-  let bulkOctoPrintUpdateButton = document.getElementById("blkOctoPrintUpdate");
   if (printer?.octoPrintUpdate?.updateAvailable) {
     if (updateButton.disabled) {
       UI.doesElementNeedUpdating(false, updateButton, "disabled");
       updateButton.setAttribute("title", "You have an OctoPrint Update to install!");
-    }
-    if (bulkOctoPrintUpdateButton.disabled) {
-      bulkOctoPrintUpdateButton.disabled = false;
     }
   } else {
     if (!updateButton.disabled) {
       UI.doesElementNeedUpdating(true, updateButton, "disabled");
       updateButton.setAttribute("title", "No OctoPrint updates available!");
     }
-    if (!bulkOctoPrintUpdateButton.disabled) {
-      bulkOctoPrintUpdateButton.disabled = true;
-    }
   }
 }
 
 function checkForOctoPrintPluginUpdates(printer) {
   let updatePluginButton = document.getElementById(`octoprintPluginUpdate-${printer._id}`);
-  let bulkPluginUpdateButton = document.getElementById("blkUpdatePluginsBtn");
   if (printer.octoPrintPluginUpdates && printer.octoPrintPluginUpdates.length > 0) {
     if (updatePluginButton.disabled) {
       updatePluginButton.disabled = false;
       updatePluginButton.title = "You have OctoPrint plugin updates to install!";
     }
-    if (bulkPluginUpdateButton.disabled) {
-      bulkPluginUpdateButton.disabled = false;
-    }
   } else {
     if (!updatePluginButton.disabled) {
       updatePluginButton.disabled = true;
       updatePluginButton.title = "No OctoPrint plugin updates available!";
-    }
-    if (!bulkPluginUpdateButton.disabled) {
-      bulkPluginUpdateButton.disabled = true;
     }
   }
 }